Backport item cheating permission checks from 1.12.2. #1226
