+++ gradle.properties
@@ -1,8 +1,8 @@
 mcversion=1.10.2
-forgeversion=12.18.3.2215
+forgeversion=12.18.3.2254
 mcp_mappings=snapshot_20161111
 curse_project_id=238222
 
 version_major=3
 version_minor=14
-version_patch=7
+version_patch=8
