+++ src/main/java/mezz/jei/network/packets/PacketCheatPermission.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 import mezz.jei.config.Config;
 import mezz.jei.network.IPacketId;
 import mezz.jei.network.PacketIdClient;
-import mezz.jei.util.CommandUtil;
+import mezz.jei.util.CommandUtilServer;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.text.TextFormatting;
@@ -32,8 +32,8 @@ public class PacketCheatPermission extends PacketJei {
 		public void readPacketData(PacketBuffer buf, EntityPlayer player) throws IOException {
 			boolean hasPermission = buf.readBoolean();
 			if (!hasPermission && Config.isCheatItemsEnabled()) {
-				CommandUtil.writeChatMessage(player, "jei.chat.error.no.cheat.permission.1", TextFormatting.RED);
-				CommandUtil.writeChatMessage(player, "jei.chat.error.no.cheat.permission.2", TextFormatting.RED);
+				CommandUtilServer.writeChatMessage(player, "jei.chat.error.no.cheat.permission.1", TextFormatting.RED);
+				CommandUtilServer.writeChatMessage(player, "jei.chat.error.no.cheat.permission.2", TextFormatting.RED);
 				Config.setCheatItemsEnabled(false);
 				player.closeScreen();
 			}
